# Optimize get_player_position